refactor(model-objects): move `SavedEpisode` to `miscellaneous`
